refactor: split benchmark period metrics
